+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - [10997](https://github.com/vegaprotocol/vega/issues/10997) - Handle cases where leaving opening auction triggers monitoring auction.
 - [11010](https://github.com/vegaprotocol/vega/issues/11010) - Release bond in spot market when market is terminated. 
 - [11016](https://github.com/vegaprotocol/vega/issues/11016) - Update holding account when amending order with cancel/replace on remaining funds.
+- [11038](https://github.com/vegaprotocol/vega/issues/11038) - Simplify price ranges state. 
 
 ## 0.75.0